+++ src/components/Badge/badge.test.js
@@ -1,18 +1,18 @@
 import React from 'react';
-import Badget from './badget';
+import Badge from './badge';
 import renderer from 'react-test-renderer';
 import { shallow } from 'enzyme';
 
 describe('render', () => {
   it('renders without crashing', () => {
-    const component = renderer.create(<Badget>Foo</Badget>);
+    const component = renderer.create(<Badge>Foo</Badge>);
     let tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('color parameter', () => {
     const component = renderer.create(
-      <Badget color="rgba(128, 182, 34, 0.24)">Foo</Badget>
+      <Badge color="rgba(128, 182, 34, 0.24)">Foo</Badge>
     );
     let tree = component.toJSON();
     expect(tree).toMatchSnapshot();
@@ -21,9 +21,9 @@ describe('render', () => {
   it('click action', () => {
     const spy = jest.fn();
     const component = shallow(
-      <Badget closable onClose={spy}>
+      <Badge closable onClose={spy}>
         Store 2
-      </Badget>
+      </Badge>
     );
 
     const button = component.find('button');
